Created AbstractContainerBuilderFactory
